Print cursor position with integer

+++ src/main.cpp
@@ -106,7 +106,7 @@ int main()
 
         /* Draw mouse cursor position text */
         std::stringstream cursorText;
-        cursorText << "x: " << cursorPos.x << ", y: " << cursorPos.y;
+        cursorText << "x: " << static_cast<int>(cursorPos.x) << ", y: " << static_cast<int>(cursorPos.y);
         renderText.render_text(cursorText.str(), static_cast<float>(cursorPos.x), static_cast<float>(cursorPos.y), 1.0f,
                                glm::vec3(0.0f));
